Separate CLI parsing & CRUD methods, Use thiserror
The parsing for the command line arguments was moved into it's own file
(mod). And yet another file is now accommodating all Crud methods. This
was done to clean up the main.rs file, so it's easier to implement a GUI
and other features later. Another big improvement came in the form of
better error handling by using the thiserror crate and defining custom
errors. That being said, there's still room for further refinement. This
commit is a major step towards a MVP.