Add sequential & autoincremented task id
In addition to a uuid each task now has an interger id. This makes it
easier to pick tasks on the command line. The id number starts at 1 and
is incremented for each new task added to the database using the
built-in AUTOINCREMENT command in SQLite. Using this method has a slight
flaw, but which does not impact functionality: When a task is deleted
from the database a gap will form in the numbering of tasks, so that
they are not sequential any longer.
Example:
The Database has: Task 1, Task 2, Task 3. When Task 2 is deleted, Task 3
will not get the task id 2. Instead the Database now looks like this:
Task 1, Task 3 (There's no Task with Id 2 any longer).